backend/requirements.txt
# Standard dependencies
asgiref
backports.zoneinfo;python_version<"3.9"
django
python-dotenv
djangorestframework
django-cors-headers

# PlanetScale dependencies
pymysql
django-mysql
mysqlclient
sqlparse
pytz

# auth dependencies
Authlib

# testing dependencies
pytest
pytest-django
pytest-cov
